requirement.txt
torch>=2.0.0         
torchvision>=0.15.1 
opencv-python>=4.11.0.86  
numpy>=1.24.0        
matplotlib>=3.10.1   
scikit-learn>=1.2.2  
tqdm>=4.67.1        
Pillow>=11.1.0       
torchsummary>=1.5.1  
scikit-image>=0.25.2 